requirements.txt
asgiref>=3.7.2
pytz>=2023.3
PyJWT>=2.8.0
sqlparse>=0.4.4
tzdata>=2023.3
Django>=4.2.0
djangorestframework>=3.14.0
django-environ>=0.9.0
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.8
django-cors-headers>=4.0.0

